refractor: move to styles.tsx convention
