+++ src/content/home.tsx
@@ -1,6 +1,6 @@
 import { Box, Typography } from '@mui/material'
 import { Helmet } from 'react-helmet-async'
-import Footer from '../../components/Footer'
+import Footer from '../components/Footer'
 
 function HomeContent() {
   return (
