+++ src/layouts/BaseLayout/index.tsx
@@ -1,7 +1,6 @@
 import PropTypes from 'prop-types'
 import { FC, ReactNode } from 'react'
 import { Outlet } from 'react-router-dom'
-
 import { Box } from '@mui/material'
 
 interface BaseLayoutProps {
