+++ src/theme/base.ts
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { Theme } from '@mui/material'
-import { NESSTheme } from './schemes/NESSTheme'
+import { NESSTheme } from './NESSTheme'
 
 export function themeCreator(theme: string): Theme {
   return themeMap[theme]
